pyproject.toml
[project]
name = "ddos_project"
version = "0.1.0"
description = ""
authors = [
    {name = "SAHIL PARASHAR",email = "sahilparashar361@gmail.com"}
]
package-mode = false
requires-python = ">=3.11"
dependencies = [
    "pandas (>=2.3.0,<3.0.0)",
    "numpy (>=2.3.0,<3.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "scikit-learn (>=1.7.0,<2.0.0)",
    "ipykernel (>=6.29.5,<7.0.0)",
    "pyarrow (>=20.0.0,<21.0.0)",
    "joblib (>=1.5.1,<2.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "kafka-python (>=2.2.11,<3.0.0)",
    "elasticsearch (>=8.0.0,<9.0.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "requests (>=2.32.4,<3.0.0)",
    "fluent-logger (>=0.10.0,<1.0.0)",
    "aiohttp (>=3.9.0,<4.0.0)",
    "psutil (>=5.9.0,<6.0.0)",
    "pydantic (>=2.0.0,<3.0.0)",
    "dpkt (>=1.9.8,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
